feat: Established new way to process imports
